// src/types/cocktails.ts
export interface IRawCocktailDrink {
  idDrink: string
  strDrink: string
  strDrinkAlternate?: string | null
  strTags?: string | null
  strVideo?: string | null
  strCategory: string
  strIBA?: string | null
  strAlcoholic: 'Alcoholic' | 'Non alcoholic' | 'Optional alcohol'
  strGlass: string
  strInstructions: string
  strInstructionsES?: string | null
  strInstructionsDE?: string | null
  strInstructionsFR?: string | null
  strInstructionsIT?: string | null
  strInstructionsZH_HANS?: string | null
  strInstructionsZH_HANT?: string | null
  strDrinkThumb: string
  strImageSource?: string | null
  strImageAttribution?: string | null
  strCreativeCommonsConfirmed?: 'Yes' | 'No' | null
  dateModified?: string | null
  strIngredient1?: string | null
  strIngredient2?: string | null
  strIngredient3?: string | null
  strIngredient4?: string | null
  strIngredient5?: string | null
  strIngredient6?: string | null
  strIngredient7?: string | null
  strIngredient8?: string | null
  strIngredient9?: string | null
  strIngredient10?: string | null
  strIngredient11?: string | null
  strIngredient12?: string | null
  strIngredient13?: string | null
  strIngredient14?: string | null
  strIngredient15?: string | null
  strMeasure1?: string | null
  strMeasure2?: string | null
  strMeasure3?: string | null
  strMeasure4?: string | null
  strMeasure5?: string | null
  strMeasure6?: string | null
  strMeasure7?: string | null
  strMeasure8?: string | null
  strMeasure9?: string | null
  strMeasure10?: string | null
  strMeasure11?: string | null
  strMeasure12?: string | null
  strMeasure13?: string | null
  strMeasure14?: string | null
  strMeasure15?: string | null
}

export interface ICocktailsResponse {
  drinks: IRawCocktailDrink[]
}

export interface IngredientPair {
  ingredient: string
  measure: string | null
}

export interface Drink {
  idDrink: string
  strDrink: string
  strDrinkAlternate: string | null
  strTags: string | null
  strVideo: string | null
  strCategory: string | null
  strIBA: string | null
  strAlcoholic: string | null
  strGlass: string | null
  strInstructions: string | null
  strInstructionsES: string | null
  strInstructionsDE: string | null
  strInstructionsFR: string | null
  strInstructionsIT: string | null
  strInstructionsZH_HANS: string | null
  strInstructionsZH_HANT: string | null
  strDrinkThumb: string | null
  strImageSource: string | null
  strImageAttribution: string | null
  strCreativeCommonsConfirmed: string | null
  dateModified: string | null

  ingredients: IngredientPair[]
}